그룹생성 화면 비율, 그룹선택, 기간 컴포넌트 분리
1. 그룹화면에서 비율, 그룹선택, 기간 컴포넌트 분리작업
